Restructure: Move solution loading out of VSSolution and into the loader
